requirements.txt
Mastodon.py
Pillow
ExifRead
